# dfras-docs/dfras-api-swagger.yaml
openapi: 3.0.3
info:
  title: DFRAS API
  description: |
    Delivery Failure Root Cause Analysis System (DFRAS) API
    
    A comprehensive AI-powered platform for analyzing delivery failures and providing predictive insights for logistics operations.
    
    ## Key Features
    - Multi-source data integration (orders, fleet logs, warehouse records, customer feedback)
    - AI-powered root cause analysis using LLM (all-MiniLM-L6-v2)
    - Automated correlation engine for pattern recognition
    - Intelligent reporting with actionable recommendations
    - Predictive analytics for proactive failure prevention
    
    ## Authentication
    All endpoints require JWT authentication. Use the `/auth/login` endpoint to obtain a token.
    
    ## Sample Data
    The system includes comprehensive sample data from `third-assignment-sample-data-set`:
    - 750+ clients across multiple cities
    - 15,000+ orders with delivery status
    - 50+ warehouses with capacity data
    - 200+ drivers with performance metrics
    - External factors (weather, traffic, events)
    - Customer feedback and ratings
  version: 1.0.0
  contact:
    name: DFRAS Support
    email: support@dfras.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local Development Server
  - url: http://localhost:8002
    description: Analytics Service Direct Access
  - url: http://localhost:8001
    description: Data Service Direct Access
security:
  - bearerAuth: []
paths:
...
  /api/analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Get Dashboard Metrics
      description: Retrieve comprehensive dashboard metrics including total orders, success rates, revenue, and failure analysis
      responses:
        '200':
          description: Dashboard metrics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_orders:
                    type: integer
                    example: 15000
                  successful_orders:
                    type: integer
                    example: 12000
                  failed_orders:
                    type: integer
                    example: 2000
                  pending_orders:
                    type: integer
                    example: 1000
                  success_rate:
                    type: number
                    format: float
                    example: 80.0
                  total_revenue:
                    type: number
                    format: float
                    example: 5076410.76
                  lost_revenue:
                    type: number
                    format: float
                    example: 5244984.42
                  avg_order_value:
                    type: number
                    format: float
                    example: 2614.01
                  avg_failed_amount:
                    type: number
                    format: float
                    example: 2622.49
                  top_failure_reasons:
                    type: array
                    items:
                      type: object
                      properties:
                        reason:
                          type: string
                          example: "Stockout"
                        count:
                          type: integer
                          example: 425
                        percentage:
                          type: number
                          format: float
                          example: 21.2
                        total_amount:
                          type: number
                          format: float
                          example: 1112500.0
                        avg_amount:
                          type: number
                          format: float
                          example: 2617.65
                  orders_by_status:
                    type: array
                    items:
                      type: object
                      properties:
                        status:
                          type: string
                          example: "Delivered"
                        count:
                          type: integer
                          example: 12000
                  orders_by_state:
                    type: array
                    items:
                      type: object
                      properties:
                        state:
                          type: string
                          example: "Maharashtra"
                        count:
                          type: integer
                          example: 3000
                  orders_by_city:
                    type: array
                    items:
                      type: object
                      properties:
                        city:
                          type: string
                          example: "Mumbai"
                        count:
                          type: integer
                          example: 1500
                  daily_trends:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                          example: "2024-01-15"
                        total_orders:
                          type: integer
                          example: 500
                        successful_orders:
                          type: integer
                          example: 400
                        failed_orders:
                          type: integer
                          example: 80
                  data_source:
                    type: string
                    example: "third-assignment-sample-data-set"
                  last_updated:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
  /api/analytics/failure-analysis:
    get:
      tags:
        - Analytics
      summary: Get Failure Analysis
      description: Retrieve detailed failure analysis with patterns, correlations, and insights
      responses:
        '200':
          description: Failure analysis retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_failures:
                    type: integer
                    example: 2000
                  failure_rate:
                    type: number
                    format: float
                    example: 13.33
                  top_failure_reasons:
                    type: array
                    items:
                      type: object
                      properties:
                        reason:
                          type: string
                          example: "Stockout"
                        count:
                          type: integer
                          example: 425
                        percentage:
                          type: number
                          format: float
                          example: 21.2
                        total_amount:
                          type: number
                          format: float
                          example: 1112500.0
                        avg_amount:
                          type: number
                          format: float
                          example: 2617.65
                  city_wise_failures:
                    type: array
                    items:
                      type: object
                      properties:
                        city:
                          type: string
                          example: "Mumbai"
                        failures:
                          type: integer
                          example: 300
                        failure_rate:
                          type: number
                          format: float
                          example: 15.5
                  warehouse_failures:
                    type: array
                    items:
                      type: object
                      properties:
                        warehouse_id:
                          type: string
                          example: "WH_001"
                        failures:
                          type: integer
                          example: 150
                        failure_rate:
                          type: number
                          format: float
                          example: 12.8
                  time_patterns:
                    type: object
                    properties:
                      peak_failure_hours:
                        type: array
                        items:
                          type: integer
                          example: 14
                      peak_failure_days:
                        type: array
                        items:
                          type: string
                          example: "Monday"
                  external_factors:
                    type: object
                    properties:
                      weather_correlation:
                        type: number
                        format: float
                        example: 0.75
                      traffic_correlation:
                        type: number
                        format: float
                        example: 0.68
        '500':
          description: Internal server error
  /api/analytics/sample-data-info:
    get:
      tags:
        - Analytics
      summary: Get Sample Data Information
      description: Retrieve metadata and statistics about the sample dataset
      responses:
        '200':
          description: Sample data information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data_statistics:
                    type: object
                    properties:
                      clients:
                        type: object
                        properties:
                          total_records:
                            type: integer
                            example: 750
                          columns:
                            type: array
                            items:
                              type: string
                            example: ["client_id", "client_name", "city", "state"]
                      orders:
                        type: object
                        properties:
                          total_records:
                            type: integer
                            example: 15000
                          columns:
                            type: array
                            items:
                              type: string
                            example: ["order_id", "client_id", "status", "amount"]
                      warehouses:
                        type: object
                        properties:
                          total_records:
                            type: integer
                            example: 50
                          columns:
                            type: array
                            items:
                              type: string
                            example: ["warehouse_id", "warehouse_name", "city", "capacity"]
                      drivers:
                        type: object
                        properties:
                          total_records:
                            type: integer
                            example: 200
                          columns:
                            type: array
                            items:
                              type: string
                            example: ["driver_id", "driver_name", "phone", "city"]
                  data_source:
                    type: string
                    example: "third-assignment-sample-data-set"
                  data_path:
                    type: string
                    example: "/app/sample-data"
                  last_updated:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
  /api/analytics/sample-data:
    get:
      tags:
        - Analytics
      summary: Get Sample Data Records
      description: Retrieve actual sample data records from the dataset
      parameters:
        - name: limit
          in: query
          description: Maximum number of records to return per data type
          required: false
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
      responses:
        '200':
          description: Sample data records retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  data:
                    type: object
                    properties:
                      clients:
                        type: array
                        items:
                          type: object
                          properties:
                            client_id:
                              type: string
                              example: "CLIENT_001"
                            client_name:
                              type: string
                              example: "ABC Corp"
                            city:
                              type: string
                              example: "Mumbai"
                            state:
                              type: string
                              example: "Maharashtra"
                      orders:
                        type: array
                        items:
                          type: object
                          properties:
                            order_id:
                              type: string
                              example: "ORD_001"
                            client_id:
                              type: string
                              example: "CLIENT_001"
                            status:
                              type: string
                              example: "Delivered"
                            amount:
                              type: number
                              format: float
                              example: 2500.0
                            order_date:
                              type: string
                              format: date-time
                            failure_reason:
                              type: string
                              example: null
                  data_source:
                    type: string
                    example: "third-assignment-sample-data-set"
                  data_path:
                    type: string
                    example: "/app/sample-data"
                  limit:
                    type: integer
                    example: 100
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...